yet another significant rewrite! but more accurate
turns out the lines in 20020 aren't lines of constant bearing, because
when you take a 1x20 PNG and drag it across the map from coast to coast,
you get a line of constant slope on an equirectangular projection.

footage from chapter 9 gives us pretty good survey markers for UAB's
field, as well as a landmark to go off of (the field is about "200
yards" away from the stannard rock lighthouse). enough to tell us that a
rhumb line wasn't right.

this is now pretty close to jon's "take an image and stretch it across
the united states" approach to building the bowl game's fields, but
we're going about it from the other direction: calculating the center of
the field (_not_ the zero yard line) and rotating it about that center.
shorter fields (SDSU) and roughly north-south fields (UAB) look great.
longer east-west fields look... not so great; i'll need to figure out
how to fix grambling and washington. my hunch is that the new boundary
intersection code isn't quite right (and if i can't figure out how to
make it right, i might be able to offset the field to get it centered
right).